+++ sql/rpl_mi.h
@@ -170,7 +170,7 @@ public:
 };
 
 bool check_master_connection_name(LEX_STRING *name);
-void create_logfile_name_with_suffix(char *res_file_name, uint length,
+void create_logfile_name_with_suffix(char *res_file_name, size_t length,
                              const char *info_file, 
                              bool append,
                              LEX_STRING *suffix);
